rename action to package
