+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/commands/OptimizeTableCommand.java
@@ -28,10 +28,10 @@ import org.apache.doris.datasource.CatalogIf;
 import org.apache.doris.datasource.ExternalTable;
 import org.apache.doris.nereids.trees.expressions.Expression;
 import org.apache.doris.nereids.trees.plans.PlanType;
-import org.apache.doris.nereids.trees.plans.commands.action.OptimizeAction;
-import org.apache.doris.nereids.trees.plans.commands.action.OptimizeActionFactory;
 import org.apache.doris.nereids.trees.plans.commands.info.PartitionNamesInfo;
 import org.apache.doris.nereids.trees.plans.commands.info.TableNameInfo;
+import org.apache.doris.nereids.trees.plans.commands.optimize.OptimizeAction;
+import org.apache.doris.nereids.trees.plans.commands.optimize.OptimizeActionFactory;
 import org.apache.doris.nereids.trees.plans.visitor.PlanVisitor;
 import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.qe.StmtExecutor;
