+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/commands/optimize/BaseOptimizeAction.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.nereids.trees.plans.commands.action;
+package org.apache.doris.nereids.trees.plans.commands.optimize;
 
 import org.apache.doris.analysis.UserIdentity;
 import org.apache.doris.catalog.Env;
